% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{BetaGen}
\alias{BetaGen}
\title{Generate \code{Beta(a,b)} samples by AR method}
\usage{
BetaGen(n, a = 1, b = 1)
}
\arguments{
\item{n}{int, the number of samples.}

\item{a}{float, first parameter of beta distribution.}

\item{b}{float, second parameter of beta distribution.}
}
\value{
a vector of samples of length \code{n}.
}
\description{
Function `BetaGen` implements Acceptance-Rejection method to generate $Beta(a,b)$ samples of size `n`.
}
\examples{
\dontrun{
BetaGen(1e4)
}
}
