% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{CI_lognormal}
\alias{CI_lognormal}
\title{Monte Carlo method to obtain empirical confidence interval}
\usage{
CI_lognormal(x)
}
\arguments{
\item{x}{vector, a sample of length `n`.}
}
\value{
a confidence interval simulated by Monte Carlo method.
}
\description{
Function `CI_lognormal` implements the Monte Carlo method to obtain empirical confidence interval.
}
\examples{
\dontrun{
x <- rlnorm(50)
CI_lognormal(x)
}
}
