% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACV.R
\name{dataloader}
\alias{dataloader}
\title{prepare the kth fold data}
\usage{
dataloader(k, loc, data)
}
\arguments{
\item{k}{int, the kth fold.}

\item{loc}{vector, the index of splitting.}

\item{data}{matrix, the data to be divided.}
}
\value{
a list of \code{X_train}, \code{y_train}, \code{X_test}, \code{y_test}
}
\description{
\code{dataloader} is a function that prepares the kth training data and testing data
}
\examples{
\dontrun{
data(PC1)
n <- nrow(PC1)
K <- 5
loc <- Ksplit(K, n)
dim(dataloader(1, loc, PC1))
dim(PC1)
}
}
