% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
GibbsC(x0, y0, n)
}
\arguments{
\item{x0}{float, first component of the start point}

\item{y0}{float, second component of the start point}

\item{n}{int, the number of samples to be generated}
}
\value{
a matrix of shape n*2,
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
GibbsC(0,0,1000)
}
}
