% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{GibbsR}
\alias{GibbsR}
\title{Implement Gibbs sampler by R}
\usage{
GibbsR(x0, n)
}
\arguments{
\item{x0}{vector of length 2, a start value.}

\item{n}{int, the number of the sample to be generated.}
}
\value{
a sample of length n.
}
\description{
Function `GibbsR` implements the Gibbs sampler by R.
}
\examples{
\dontrun{
GibbsR(c(0,0), 1000)
}
}
