% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{ImpSam}
\alias{ImpSam}
\title{Implement importance sampling method}
\usage{
ImpSam(n, m = 100)
}
\arguments{
\item{n}{int, the number of times to do Monte Carlo.}

\item{m}{int, the number of times to estimate.}
}
\value{
a vector of estimation of length \code{m}.
}
\description{
Function `ImpSam` implements the importance sampling method.
}
\examples{
\dontrun{
ImpSam(300)
}
}
