% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACV.R
\name{Kfold_weight}
\alias{Kfold_weight}
\title{Calculate the weight of model averaging}
\usage{
Kfold_weight(data, K, c, method = "squareloss")
}
\arguments{
\item{data}{the data used to calculate weight.}

\item{K}{K-fold cross validation model averaging method.}

\item{c}{the threshold to be positive.}

\item{method}{optimization methods: "squareloss" only.}
}
\value{
a vector of weight.
}
\description{
\code{Kfold_weight} is a function that calculates the weight of model averaging.
}
\examples{
\dontrun{
data(PC1)
Kfold_weight(PC1, 5, 0.2)
}
}
