% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACV.R
\name{Ksplit}
\alias{Ksplit}
\title{Generate an index vector to split the data-set}
\usage{
Ksplit(K, n)
}
\arguments{
\item{K}{int, the number of folds divided.}

\item{n}{int, the number of samples.}
}
\value{
a vector of splitting index.
}
\description{
\code{Ksplit} is a function, which can generate an index vector to split the data-set to \code{K} parts.
}
\examples{
\dontrun{
K <- 5
n <- 17
Ksplit(K, n)
}
}
