% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACV.R
\name{MAtree}
\alias{MAtree}
\title{Model averaging on \code{xgboost} and \code{lightgbm}}
\usage{
MAtree(data, k = 5, m = 3, c = 0.2, method = "squareloss")
}
\arguments{
\item{data}{the data used to train and test.}

\item{k}{k-fold cross validation to evaluate the performance.}

\item{m}{m=fold cross validation model averaging method.}

\item{c}{the threshold to be positive.}

\item{method}{optimization methods: "squareloss" only.}
}
\value{
The prediction and the performance of \code{lightgbm}, \code{xgboost} and \code{MAtree}.
}
\description{
Conduct model averaging method on \code{xgboost} and \code{lightgbm} to do software defect prediction.
}
\examples{
\dontrun{
data(PC1)
str(MAtree(PC1))
}
}
