% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{MixGen}
\alias{MixGen}
\title{Generate Exponential-Gamma mixture samples}
\usage{
MixGen(n, gamma = 4, beta = 2)
}
\arguments{
\item{n}{int, the number of samples.}

\item{gamma}{float, first parameter of gamma distribution.}

\item{beta}{float, second parameter of gamma distribution.}
}
\value{
a vector of samples of length \code{n}.
}
\description{
Function `MixGen` simulates a continuous Exponential-Gamma mixture.
}
\examples{
\dontrun{
MixGen(1e4)
}
}
