% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{MonteGen}
\alias{MonteGen}
\title{Compare simple Monte Carlo and antithetic variate approach}
\usage{
MonteGen(n, antithetic = TRUE)
}
\arguments{
\item{n}{int, the number of times to do Monte Carlo simulation.}

\item{antithetic}{bool, use antithetic or not.}
}
\value{
Monte Catlo simulation result.
}
\description{
Function `MonteGen` can calculate integration of $e^x$ on $[0,1]$ with or without antithetic variate approach.
}
\examples{
\dontrun{
MonteGen(100)
}
}
