% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{ParatoGen}
\alias{ParatoGen}
\title{Generate \code{Parato(2,2)} samples by inverse transformation method}
\usage{
ParatoGen(n)
}
\arguments{
\item{n}{int, the number of samples.}
}
\value{
a vector of samples of length \code{n}.
}
\description{
Function \code{ParatoGen} implements inverse transformation method to generate `Parato(2,2)` samples of size `n`.
}
\examples{
\dontrun{
ParatoGen(1e4)
}
}
