% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACV.R
\name{perform}
\alias{perform}
\title{Evaluate the performance of the prediction}
\usage{
perform(y_true, y_pred, y_proba)
}
\arguments{
\item{y_true}{vector, the true response.}

\item{y_pred}{vector, the prediction of the response.}

\item{y_proba}{vector, the predicted probability of the response.}
}
\value{
a list of performance with \code{Accuracy}, \code{Precision}, \code{Recall}, \code{F1}, \code{AUC}.
}
\description{
\code{perform} is a function that can evaluate the performance of the prediction.
}
\examples{
\dontrun{
data(PC1)
datafull <- dataloader(1, loc, PC1)
X_test <- as.matrix(datafull$X_test)
y_test <- as.matrix(datafull$y_test)
X_train <- as.matrix(datafull$X_train)
y_train <- as.matrix(datafull$y_train)
dtrain = lgb.Dataset(X_train, label = y_train)
model <- lightgbm(data=dtrain, verbose=-1, params = list(objective="binary"))
yr <- predict(model, X_test)
yp <- c(0,1)[(yr>0.2)+1]
str(perform(y_test, yp, yr))
}
}
