\name{StatComp22087-package}
\alias{StatComp22087-package}
\alias{StatComp22087}
\docType{package}
\title{
  MAtree algorithm implementation and homework for 'Statistical Computing' course
}
\description{
  MAtree algorithm and several R and Rcpp functions are included in this package for illustration. 
}
\author{
Yixi Dong

Maintainer: Yixi Dong <kasha@mail.ustc.edu.cn>
}
\references{
  Xinyu Zhang, Chu-An Liu, Model averaging prediction by K-fold cross-validation, Journal of Econometrics, 2022, ISSN 0304-4076.
}
\keyword{ package }
