% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{StrImpSam}
\alias{StrImpSam}
\title{Implement stratified importance sampling method}
\usage{
StrImpSam(n, m = 100)
}
\arguments{
\item{n}{int, the number of times to do Monte Carlo.}

\item{m}{int, the number of times to estimate.}
}
\value{
a vector of estimation of length \code{m}.
}
\description{
Function `StrImpSam` implements the stratified importance sampling method.
}
\examples{
\dontrun{
StrImpSam(300)
}
}
